move get_rank
