+++ python/oneflow/__init__.py
@@ -97,7 +97,7 @@ def _SyncOnMasterFn():
         return
     if oneflow.framework.distribute.is_multi_client():
         oneflow._oneflow_internal.eager.multi_client.Sync()
-    elif oneflow.framework.distribute.get_rank() == 0:
+    elif oneflow.env.get_rank() == 0:
         oneflow._oneflow_internal.eager.single_client.Sync()
