+++ python/oneflow/distributed/__init__.py
@@ -15,7 +15,7 @@ limitations under the License.
 """
 from oneflow.framework.distribute import (
     # get_local_rank,
-    get_rank,
+    # get_rank,
     get_world_size,
     is_multi_client,
 )
