+++ python/oneflow/nn/graph/block.py
@@ -19,7 +19,7 @@ from typing import Iterator, Optional, Set, Union
 
 import oneflow._oneflow_internal
 import oneflow.framework.graph_build_util as graph_build_util
-from oneflow.framework.distribute import get_rank
+from oneflow.env import get_rank
 from oneflow.framework.tensor import Tensor, TensorTuple
 from oneflow.nn.module import Module
 from oneflow.nn.parameter import Parameter
