+++ python/oneflow/utils/data/distributed.py
@@ -88,12 +88,12 @@ class DistributedSampler(Sampler[T_co]):
         if num_replicas is None:
             num_replicas = dist.get_world_size()
         if rank is None:
-            rank = dist.get_rank()
+            rank = flow.env.get_rank()
         print(
             "dist.get_world_size() >>>>> ",
             dist.get_world_size(),
-            "dist.get_rank() >>>>>",
-            dist.get_rank(),
+            "flow.env.get_rank() >>>>>",
+            flow.env.get_rank(),
         )
         if rank >= num_replicas or rank < 0:
             raise ValueError(
